# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Bravado Company Ltd Payroll API
  version: 1.0.0
  description: REST API for the Bravado Company Ltd Payroll Management System.
paths:
...
  /api/self/payslips/:
    get:
      operationId: self_payslips_list
      tags:
      - self
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Payslip'
          description: ''
...
components:
  schemas:
...
    Payslip:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        item:
          type: integer
        payslip_number:
          type: string
          maxLength: 40
        employee_number:
          type: string
          readOnly: true
        employee_name:
          type: string
          readOnly: true
        department_name:
          type: string
          readOnly: true
        period:
          type: integer
          readOnly: true
        period_name:
          type: string
          readOnly: true
        period_start:
          type: string
          format: date
          readOnly: true
        gross_earnings:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        total_deductions:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        net_salary:
          type: string
          format: decimal
          pattern: ^-?\d{0,14}(?:\.\d{0,2})?$
          readOnly: true
        generated_at:
          type: string
          format: date-time
          readOnly: true
        generated_by_name:
          type: string
          readOnly: true
        download_count:
          type: integer
          maximum: 2147483647
          minimum: 0
      required:
      - department_name
      - employee_name
      - employee_number
      - generated_at
      - generated_by_name
      - gross_earnings
      - id
      - item
      - net_salary
      - payslip_number
      - period
      - period_name
      - period_start
      - total_deductions
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT